refactor: consolidate endpoint tests into single api test module
