--- tests/test_health.py
@@ -1,10 +1,0 @@
-from fastapi.testclient import TestClient
-from governance_analysis_engine.main import app
-
-
-def test_health_endpoint_returns_ok() -> None:
-    client = TestClient(app)
-    response = client.get("/health")
-
-    assert response.status_code == 200
-    assert response.json() == {"status": "ok"}
